refactor(lab-l9yv.18): split runner.rs into host_io.rs, stages.rs, runner.rs
Extract HostIo trait + SshHostIo into host_io.rs (~100L), move the four
stage functions (preflight, transfer_and_install, restart, verify) plus
their outcome types and helpers into stages.rs (~450L), and slim runner.rs
down to orchestration, DefaultRunner, test_support, and the NoopRunner
fallback (~600L).

Stage tests move to stages.rs (referencing runner::test_support::RecordingIo).
HostIo is re-exported from runner.rs for backward compat with the integration
test import path (tests/deploy_runner.rs). Add pub mod host_io; and
pub mod stages; to dispatch/deploy.rs.
